first, non working approach of place
The code should place the object at the same spot on the table where it was before the pick.

The problem is that the position constraints are violated for the eef link "s_model_tool0".
We suppose that the cause of this error is that place uses the parent link of the eef instead of the eef itself at some point.

Important: the pose we set for the place location is the pose for the object NOT for the eef!