new pick place loop demo, NOT TESTED!
package is renamed

The new demo should pick the from the object which is detected from the camera and place it somewhere on the table.
In the moment just a pick loop is implemented, the object will be pickt, the arm will move to the start grap position and then the object will be released.

This demo is not tested yet!