Working version of pick loop
We tested the pick loop on the ur5.

Main errors were in the header frames for the object and the grasp pose.
New function to remove the attached object from the planning scene.

The loop picks up the object on the table and releases it after 5
seconds. Then moves back to the start point and starts over.
The position of the object is detected from the camera, so we can try
out different positions.