PlanGrasps service server
This is the server implementation for the new PlanGrasps service.
The server provides the same functionality from the old service,
the difference is that we create a moveit_msgs::Grasp instead of the old one
from manipulation_msgs.

The advantige is that we can set a time_from_start here, which we also made.
The values for the duration were just set and not messuered.